All the code is concentrated in one file
Separate the CLI logic in a separate file ('main.rs') and split the
code into three separate modules:

  * Method - a list of RPC calls
  * Request - process RPC calls
  * Serialize - serialize & deserialize JSON for RPC calls
