+++ rustfmt.toml
@@ -1,2 +1,1 @@
-tab_spaces = 4
-newline_style = "Unix"
+# max_width = 120
