+++ src/method.rs
@@ -1,0 +1,9 @@
+use jsonrpc_core::Params;
+
+pub enum Method<'a> {
+    ClientVersion,
+    EthSyncing,
+    EthBlockNumber,
+    EthAccounts,
+    EthGetBalance(&'a Params),
+}
